+++ src/main/java/com/sprint/mission/discodeit/security/DiscodeitUserDetailsService.java
@@ -1,6 +1,5 @@
-package com.sprint.mission.discodeit.service.basic;
+package com.sprint.mission.discodeit.security;
 
-import com.sprint.mission.discodeit.entity.DiscodeitUserDetails;
 import com.sprint.mission.discodeit.entity.UserEntity;
 import com.sprint.mission.discodeit.exception.user.NoSuchUserException;
 import com.sprint.mission.discodeit.mapper.UserEntityMapper;
